test: keep functional recording support boundary
